Upgrade to Docusaurus v2.0
